Create Interface for Camera, move get world point into gameservice to prevent circular dependency
