+++ Assets/Scripts/GamePlay/Services/GamePlayService.cs
@@ -1,28 +1,83 @@
+using Source.GamePlay.Services.Interfaces;
 using Source.GamePlay.Services.Units;
-using Source.Shared.Services;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
+using UnityEngine.EventSystems;
+using UnityEngine.InputSystem;
 
 namespace Source.GamePlay.Services
 {
+    public class ContactDto
+    {
+        public bool HitGameObject = false;
+        public Vector3? Point = null;
+        public GameObject GameObject = null;
+    }
+
     public class GamePlayService
     {
-        private InputService InputService;
+        private ICameraService CameraService;
         private UnitService UnitService;
 
-        public void InjectDependencies(InputService inputService, UnitService unitService)
+        public void InjectDependencies(ICameraService cameraService, UnitService unitService)
         {
-            InputService = inputService;
+            CameraService = cameraService;
             UnitService = unitService;
         }
 
         public void PrimaryClickEvent()
         {
-            ContactDto contact = InputService?.GetMouseWorldPoint();
+            ContactDto contact = GetMouseWorldPoint();
             if (contact != null)
             {
                 Debug.Log($"Hit GameObject: {contact.HitGameObject} | WorldPoint: {contact.Point} | GameObject: {contact.GameObject?.name}");
             }
+        }
+
+        private ContactDto GetMouseWorldPoint()
+        {
+            ContactDto contact = new ContactDto();
+            Vector2 mousePosition = Mouse.current.position.value;
+            if (CameraService == null)
+            {
+                return contact;
+            }
+
+            if (MouseIsOverUI(mousePosition))
+            {
+                return contact;
+            }
+
+            Ray ray;
+            if (!CameraService.ScreenToWorldPoint(mousePosition, out ray))
+            {
+                return contact;
+            }
+
+            RaycastHit hit;
+            int layerMaskToHit = LayerMask.GetMask("Default");
+            if (Physics.Raycast(ray, out hit, Mathf.Infinity, layerMaskToHit))
+            {
+                contact.HitGameObject = true;
+                contact.Point = hit.point;
+                contact.GameObject = hit.collider.gameObject;
+            }
+
+            return contact;
+        }
+
+        private bool MouseIsOverUI(Vector2 mousePosition)
+        {
+            int uiLayer = LayerMask.NameToLayer("UI");
+
+            PointerEventData eventData = new PointerEventData(EventSystem.current);
+            eventData.position = Mouse.current.position.value;
+            List<RaycastResult> raycastResults = new List<RaycastResult>();
+            EventSystem.current.RaycastAll(eventData, raycastResults);
+
+            return raycastResults.Where(r => r.gameObject.layer == uiLayer).Count() > 0;
         }
 
         public Action PrimaryHoldEvent = () => { };
