+++ Assets/Scripts/GamePlay/Services/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1be8ada35a950b24c96d1ef784f66001
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
