+++ Assets/Scripts/GamePlay/Services/Interfaces/ICameraService.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace Source.GamePlay.Services.Interfaces
+{
+    public interface ICameraService
+    {
+        public bool ScreenToWorldPoint(Vector2 mousePosition, out Ray ray);
+        public void OnMove(Vector2 direction);
+    }
+}
