+++ Assets/Scripts/GamePlay/Services/Interfaces/ICameraService.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f3886a2935e03e54f983c1dabe5ef875
